+++ cpp/binary/pymod_modmesh/module.cpp
@@ -22,56 +22,6 @@
 #include <modmesh/python/python.hpp> // Must be the first include.
 #include <modmesh/python/wrapper/modmesh/modmesh.hpp>
 
-#include <pybind11/numpy.h>
-#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
-#include <numpy/arrayobject.h>
-
-namespace modmesh
-{
-
-namespace python
-{
-
-namespace detail
-{
-
-static void import_numpy()
-{
-    auto local_import_numpy = []()
-    {
-        import_array2("cannot import numpy", false); // or numpy c api segfault.
-        return true;
-    };
-    if (!local_import_numpy())
-    {
-        throw pybind11::error_already_set();
-    }
-}
-
-} /* end namespace detail */
-
-struct modmesh_pymod_tag;
-
-static void initialize_modmesh(pybind11::module & mod)
-{
-    auto initialize_impl = [](pybind11::module & mod)
-    {
-        detail::import_numpy();
-
-        wrap_profile(mod);
-        wrap_ConcreteBuffer(mod);
-        wrap_SimpleArray(mod);
-        wrap_StaticGrid(mod);
-        wrap_StaticMesh(mod);
-    };
-
-    OneTimeInitializer<modmesh_pymod_tag>::me()(mod, initialize_impl);
-}
-
-} /* end namespace python */
-
-} /* end namespace modmesh */
-
 PYBIND11_MODULE(_modmesh, mod) // NOLINT
 {
     modmesh::python::initialize_modmesh(mod);
