+++ cpp/modmesh/python/CMakeLists.txt
@@ -13,7 +13,7 @@ set(MODMESH_PYTHON_HEADERS
 )
 
 set(MODMESH_PYTHON_SOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/python.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/common.cpp
     CACHE FILEPATH "" FORCE
 )
