+++ cpp/modmesh/python/common.hpp
@@ -31,6 +31,7 @@
 #include <pybind11/pybind11.h> // Must be the first include.
 #include <pybind11/attr.h>
 #include <pybind11/numpy.h>
+#include <pybind11/embed.h>
 
 #include <atomic>
 
@@ -44,6 +45,7 @@
 
 namespace modmesh
 {
+
 namespace python
 {
 
@@ -384,6 +386,92 @@ private:
     class_ m_cls;
 
 }; /* end class WrapBase */
+
+void import_numpy();
+
+#pragma GCC diagnostic push
+// Suppress the warning "greater visibility than the type of its field"
+#pragma GCC diagnostic ignored "-Wattributes"
+/**
+ * Take a pybind11 module and an initializing function and only run the
+ * initializing function once.
+ */
+template <typename T>
+class OneTimeInitializer
+{
+
+public:
+
+    OneTimeInitializer(OneTimeInitializer const &) = delete;
+    OneTimeInitializer(OneTimeInitializer &&) = delete;
+    OneTimeInitializer & operator=(OneTimeInitializer const &) = delete;
+    OneTimeInitializer & operator=(OneTimeInitializer &&) = delete;
+    ~OneTimeInitializer() = default;
+
+#ifdef IAMWWRONG
+    static OneTimeInitializer<T> & me()
+    {
+        static OneTimeInitializer<T> instance;
+        return instance;
+    }
+#else // IAMWWRONG
+    // Do not implement this function as a template. It should use a specialization in only one compilation unit.
+    static OneTimeInitializer<T> & me();
+#endif // IAMWWRONG
+
+    OneTimeInitializer<T> & operator()(
+        pybind11::module & mod, std::function<void(pybind11::module &)> const & initializer)
+    {
+        if (!initialized())
+        {
+            m_mod = &mod;
+            m_initializer = initializer;
+            m_initializer(*m_mod);
+        }
+        m_initialized = true;
+        return *this;
+    }
+
+    pybind11::module const & mod() const { return *m_mod; }
+    pybind11::module & mod() { return *m_mod; }
+
+    bool initialized() const { return m_initialized && nullptr != m_mod; }
+
+private:
+
+    OneTimeInitializer() = default;
+
+    bool m_initialized = false;
+    pybind11::module * m_mod = nullptr;
+    std::function<void(pybind11::module &)> m_initializer;
+
+}; /* end class OneTimeInitializer */
+#pragma GCC diagnostic pop
+
+class MODMESH_PYTHON_WRAPPER_VISIBILITY Interpreter
+{
+
+public:
+
+    static Interpreter & instance();
+
+    Interpreter(Interpreter const &) = delete;
+    Interpreter(Interpreter &&) = delete;
+    Interpreter & operator=(Interpreter const &) = delete;
+    Interpreter & operator=(Interpreter &&) = delete;
+    ~Interpreter();
+
+    void preload_module(std::string const & name);
+    void preload_modules(std::vector<std::string> const & names);
+
+private:
+
+    Interpreter();
+    void setup_path();
+
+    pybind11::scoped_interpreter * m_interpreter = nullptr;
+
+}; /* end class Interpreter */
 
 } /* end namespace python */
