+++ cpp/modmesh/python/python.hpp
@@ -28,100 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <pybind11/pybind11.h> // Must be the first include.
-#include <pybind11/stl.h>
-#include <pybind11/embed.h>
-
+#include <modmesh/python/common.hpp> // Must be the first include.
 #include <modmesh/modmesh.hpp>
-#include <modmesh/python/common.hpp>
-
-namespace modmesh
-{
-
-namespace python
-{
-
-#pragma GCC diagnostic push
-// Suppress the warning "greater visibility than the type of its field"
-#pragma GCC diagnostic ignored "-Wattributes"
-/**
- * Take a pybind11 module and an initializing function and only run the
- * initializing function once.
- */
-template <typename T>
-class OneTimeInitializer
-{
-
-public:
-
-    OneTimeInitializer(OneTimeInitializer const &) = delete;
-    OneTimeInitializer(OneTimeInitializer &&) = delete;
-    OneTimeInitializer & operator=(OneTimeInitializer const &) = delete;
-    OneTimeInitializer & operator=(OneTimeInitializer &&) = delete;
-    ~OneTimeInitializer() = default;
-
-    static OneTimeInitializer<T> & me()
-    {
-        static OneTimeInitializer<T> instance;
-        return instance;
-    }
-
-    OneTimeInitializer<T> & operator()(
-        pybind11::module & mod, std::function<void(pybind11::module &)> const & initializer)
-    {
-        if (!initialized())
-        {
-            m_mod = &mod;
-            m_initializer = initializer;
-            m_initializer(*m_mod);
-        }
-        m_initialized = true;
-        return *this;
-    }
-
-    pybind11::module const & mod() const { return *m_mod; }
-    pybind11::module & mod() { return *m_mod; }
-
-    bool initialized() const { return m_initialized && nullptr != m_mod; }
-
-private:
-
-    OneTimeInitializer() = default;
-
-    bool m_initialized = false;
-    pybind11::module * m_mod = nullptr;
-    std::function<void(pybind11::module &)> m_initializer;
-
-}; /* end class OneTimeInitializer */
-#pragma GCC diagnostic pop
-
-class MODMESH_PYTHON_WRAPPER_VISIBILITY Interpreter
-{
-
-public:
-
-    static Interpreter & instance();
-
-    Interpreter(Interpreter const &) = delete;
-    Interpreter(Interpreter &&) = delete;
-    Interpreter & operator=(Interpreter const &) = delete;
-    Interpreter & operator=(Interpreter &&) = delete;
-    ~Interpreter();
-
-    void preload_module(std::string const & name);
-    void preload_modules(std::vector<std::string> const & names);
-
-private:
-
-    Interpreter();
-    void setup_path();
-
-    pybind11::scoped_interpreter * m_interpreter = nullptr;
-
-}; /* end class Interpreter */
-
-} /* end namespace python */
-
-} /* end namespace modmesh */
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
