+++ cpp/modmesh/python/wrapper/modmesh/modmesh.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Yung-Yu Chen <yyc@solvcon.net>
+ * Copyright (c) 2019, Yung-Yu Chen <yyc@solvcon.net>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -28,25 +28,40 @@
 
 #include <modmesh/python/python.hpp> // Must be the first include.
 #include <modmesh/python/wrapper/modmesh/modmesh.hpp>
-#include <modmesh/python/wrapper/view/view.hpp>
-#include <modmesh/modmesh.hpp>
-#include <modmesh/view/view.hpp>
 
-PYBIND11_EMBEDDED_MODULE(_modmesh, mod) // NOLINT
+namespace modmesh
 {
-    modmesh::python::initialize_modmesh(mod);
+
+namespace python
+{
+
+struct modmesh_pymod_tag;
+
+template <>
+OneTimeInitializer<modmesh_pymod_tag> & OneTimeInitializer<modmesh_pymod_tag>::me()
+{
+    static OneTimeInitializer<modmesh_pymod_tag> instance;
+    return instance;
 }
 
-PYBIND11_EMBEDDED_MODULE(_modmesh_view, mod) // NOLINT
+void initialize_modmesh(pybind11::module & mod)
 {
-    modmesh::python::initialize_view(mod);
+    auto initialize_impl = [](pybind11::module & mod)
+    {
+        import_numpy();
+
+        wrap_profile(mod);
+        wrap_ConcreteBuffer(mod);
+        wrap_SimpleArray(mod);
+        wrap_StaticGrid(mod);
+        wrap_StaticMesh(mod);
+    };
+
+    OneTimeInitializer<modmesh_pymod_tag>::me()(mod, initialize_impl);
 }
 
-int main(int argc, char ** argv)
-{
-    modmesh::RApplication app(argc, argv);
-    app.main()->resize(1000, 600);
-    return app.exec();
-}
+} /* end namespace python */
 
-// vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
+} /* end namespace modmesh */
+
+// vim: set ff=unix fenc=utf8 nobomb et sw=4 ts=4 sts=4:
