+++ cpp/modmesh/python/wrapper/view/CMakeLists.txt
@@ -9,6 +9,7 @@ set(MODMESH_PYMOD_VIEW_HEADERS
 )
 
 set(MODMESH_PYMOD_VIEW_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/view.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/wrap_view.cpp
     CACHE FILEPATH "" FORCE
 )
