+++ cpp/modmesh/python/wrapper/view/view.cpp
@@ -26,27 +26,39 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/python/python.hpp> // Must be the first include.
-#include <modmesh/python/wrapper/modmesh/modmesh.hpp>
-#include <modmesh/python/wrapper/view/view.hpp>
+#include <modmesh/python/wrapper/view/view.hpp> // Must be the first include.
+#include <pybind11/stl.h>
+
 #include <modmesh/modmesh.hpp>
-#include <modmesh/view/view.hpp>
+#include <modmesh/python/common.hpp>
 
-PYBIND11_EMBEDDED_MODULE(_modmesh, mod) // NOLINT
+namespace modmesh
 {
-    modmesh::python::initialize_modmesh(mod);
+
+namespace python
+{
+
+struct view_pymod_tag;
+
+template <>
+OneTimeInitializer<view_pymod_tag> & OneTimeInitializer<view_pymod_tag>::me()
+{
+    static OneTimeInitializer<view_pymod_tag> instance;
+    return instance;
 }
 
-PYBIND11_EMBEDDED_MODULE(_modmesh_view, mod) // NOLINT
+void initialize_view(pybind11::module & mod)
 {
-    modmesh::python::initialize_view(mod);
+    auto initialize_impl = [](pybind11::module & mod)
+    {
+        wrap_view(mod);
+    };
+
+    OneTimeInitializer<view_pymod_tag>::me()(mod, initialize_impl);
 }
 
-int main(int argc, char ** argv)
-{
-    modmesh::RApplication app(argc, argv);
-    app.main()->resize(1000, 600);
-    return app.exec();
-}
+} /* end namespace python */
+
+} /* end namespace modmesh */
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
